Move MirrorSource class to own file
