+++ src/BaGetter.Core/Configuration/MirrorOptions.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
 
 namespace BaGetter.Core;
 
@@ -77,30 +76,5 @@ public class MirrorOptions : IValidatableObject
             return;
 
         Sources = [new MirrorSource { PackageSource = PackageSource, Legacy = Legacy }]; 
-    }
-}
-
-public class MirrorSource
-{
-    public Uri PackageSource { get; set; }
-    public bool Legacy { get; set; }
-
-    public override bool Equals(object obj)
-    {
-        // Check for null and compare run-time types.
-        if (obj is null || !GetType().Equals(obj.GetType()))
-        {
-            return false;
-        }
-        else
-        {
-            var ms = (MirrorSource)obj;
-            return (PackageSource.Equals(ms.PackageSource)) && (Legacy == ms.Legacy);
-        }
-    }
-
-    public override int GetHashCode()
-    {
-        return HashCode.Combine(PackageSource, Legacy);
     }
 }
